+++ init.rb
@@ -25,7 +25,7 @@ Redmine::Plugin.register :redmine_reporter_dashboards do
   author 'Jan Catrysse'
   description 'Dashboard extension for the Redmine Reporter plugin, adding project dashboards, ' \
               'SQL-based issue statistics and Liquid aggregation tags.'
-  version '0.2.0'
+  version '0.3.0'
   url 'https://github.com/jcatrysse/redmine_reporter_dashboards'
   author_url 'https://github.com/jcatrysse'